Added bookmark functionality

+++ db/mongo.py
@@ -12,7 +12,7 @@ import hashlib
 def check_if_present():
     pass
 
-def dataframe_from_collection(current_collection, collection_unique_id, collection_format={}):
+def dataframe_from_collection(current_collection, collection_unique_id, collection_format={}, curr_collection_schema={}):
     '''
         Converts the unstructed database documents to structured pandas dataframe
     '''
@@ -20,27 +20,34 @@ def dataframe_from_collection(current_collection, collection_unique_id, collecti
     docu_update = []
     count = 0
     total_len = current_collection.count_documents({})
-    logger.info("Found " + str(total_len) + " documents.")
+    logger.info("Total " + str(total_len) + " documents present in collection.")
     client_encr = MongoClient(encryption_store['url'], tlsCAFile=certifi.where())
     db_encr = client_encr[encryption_store['db_name']]
     collection_encr = db_encr[encryption_store['collection_name']]
-
-    for document in current_collection.find():
+    collection_documents = current_collection.find()
+    if(curr_collection_schema['bookmark']):
+        if('last_run_cron_job' not in curr_collection_schema.keys()):
+            curr_collection_schema['last_run_cron_job'] = datetime.datetime(1602, 8, 20)
+        collection_documents = current_collection.find({curr_collection_schema['bookmark']: {"$gt": curr_collection_schema['last_run_cron_job']}})
+        curr_collection_schema['last_run_cron_job'] = datetime.datetime.now()
+    logger.info("Fetched all candidate documents from collection.")
+    for document in collection_documents:
         document_to_be_encrypted = document.copy()
         document_to_be_encrypted['_id'] = str(document_to_be_encrypted['_id'])
         encr = {
             'collection': collection_unique_id,
             'map_id': document['_id'],
-            'document_sha': hashlib.sha256(json.dumps(document_to_be_encrypted).encode()).hexdigest()
+            'document_sha': hashlib.sha256(json.dumps(document_to_be_encrypted, default=str).encode()).hexdigest()
         }
         updation = False
-        previous_records = collection_encr.find_one({'collection': collection_unique_id, 'map_id': document['_id']})
-        if(previous_records):
-            if(previous_records['document_sha'] == encr['document_sha']):
-                continue     # No updation in this record
-            else:
-                updation = True     # This record has been updated
-                collection_encr.delete_one({'collection': collection_unique_id, 'map_id': document['_id']})
+        if(not curr_collection_schema['bookmark']):
+            previous_records = collection_encr.find_one({'collection': collection_unique_id, 'map_id': document['_id']})
+            if(previous_records):
+                if(previous_records['document_sha'] == encr['document_sha']):
+                    continue     # No updation in this record
+                else:
+                    updation = True     # This record has been updated
+                    collection_encr.delete_one({'collection': collection_unique_id, 'map_id': document['_id']})
         
         collection_encr.insert_one(encr)
         document['parquet_format_date_year'] = (document['_id'].generation_time - datetime.timedelta(hours=5, minutes=30)).year
@@ -61,7 +68,7 @@ def dataframe_from_collection(current_collection, collection_unique_id, collecti
             docu.append(document)
         else:
             docu_update.append(document)
-        if(count % 10000 == 0):
+        if(count % 5 == 0):
             logger.info(str(count)+ " documents fetched ... " + str(int(count*100/total_len)) + " %")
     
     logger.info(str(count)+ " documents fetched ... " + str(int(count*100/total_len)) + " %")
@@ -86,7 +93,7 @@ def process_data_from_source(database_, collection_name = []):
     for curr_collection in collection_name:
         if('fields' not in curr_collection.keys()):
             curr_collection['fields'] = {}
-        df_collection, df_collection_update = dataframe_from_collection(database_[curr_collection['collection_name']], collection_unique_id=curr_collection['collection_unique_id'], collection_format=curr_collection['fields'])
+        df_collection, df_collection_update = dataframe_from_collection(database_[curr_collection['collection_name']], collection_unique_id=curr_collection['collection_unique_id'], collection_format=curr_collection['fields'], curr_collection_schema = curr_collection)
         df_list.append({'collection_name': curr_collection['collection_name'], 'df_collection': df_collection, 'df_collection_update': df_collection_update})
     return df_list
